+++ src/main/java/com/alkemy/ong/model/Comment.java
@@ -1,5 +1,6 @@
 package com.alkemy.ong.model;
 
+import com.alkemy.ong.security.model.User;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
