--- src/main/java/com/alkemy/ong/model/Role.java
@@ -1,39 +1,0 @@
-package com.alkemy.ong.model;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
-import java.util.Date;
-
-@Entity
-@Table(name = "roles")
-@Getter
-@Setter
-public class Role {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    @NotNull(message = "Name may not be null")
-    private String name;
-
-    private String description;
-
-    @Column(name = "creation_date")
-    @Temporal(TemporalType.TIMESTAMP)
-    @CreationTimestamp
-    private Date creationDate;
-
-    @Column(name = "last_update")
-    @Temporal(TemporalType.TIMESTAMP)
-    @UpdateTimestamp
-    private Date lastUpdate;
-
-
-}
