+++ src/main/java/com/alkemy/ong/repository/NewsRepository.java
@@ -4,6 +4,12 @@ import com.alkemy.ong.model.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+import java.util.Optional;
+
 @Repository
 public interface NewsRepository extends JpaRepository<News, Long> {
+    boolean existById(Long id);
+
+
 }
