--- src/main/java/com/alkemy/ong/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.alkemy.ong.repository;
-
-import com.alkemy.ong.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    boolean existsByEmail (String email);
-    Optional<User> findByEmail(String email);
-
-}
