+++ src/main/java/com/alkemy/ong/security/dto/UserGetDto.java
@@ -1,6 +1,5 @@
 package com.alkemy.ong.security.dto;
 
-import com.alkemy.ong.model.Role;
 import lombok.Getter;
 import lombok.Setter;
