+++ src/main/java/com/alkemy/ong/security/dto/UserPostDto.java
@@ -3,7 +3,7 @@ package com.alkemy.ong.security.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Column;
+
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
