+++ src/main/java/com/alkemy/ong/security/filter/JwtRequestFilter.java
@@ -1,0 +1,4 @@
+package com.alkemy.ong.security.filter;
+
+public class JwtRequestFilter {
+}
