+++ src/main/java/com/alkemy/ong/security/repository/RoleRepository.java
@@ -1,13 +1,14 @@
-package com.alkemy.ong.repository;
+package com.alkemy.ong.security.repository;
 
 
-import com.alkemy.ong.model.Role;
+
+import com.alkemy.ong.security.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
+
 
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
-    Optional<Role> findByName (String name);
+   Role findByName (String name);
 }
