+++ src/main/resources/messages.properties
@@ -1,7 +1,7 @@
 # messages.properties
 credencials.incorrect="Incorrect username or password"
 
-user.NotFound="Username not found"
+email.notfound="User not found with the given email."
 
 test.one=  ESTO es una PRUEBA
